Share the records row and inset it like the set cells
Extract PersonalBestRow (the personal-record secondary tile shared by the Summary records tile and the workout-detail records tile) into SharedUI/Views, alongside its value helpers, so both surfaces provably use one component.

Inset the rows by CELL_PADDING/2 from the card border (horizontally and at the bottom) to match how WorkoutSetCell sits inside WorkoutSetGroupCell; the header keeps the standard CELL_PADDING inset.
